orchestrator/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis / Queue
redis==5.0.4
rq==1.16.2

# Vector DB
chromadb==0.5.0

# Object Store
boto3==1.34.100
aiobotocore==2.13.0

# LLM Adapters
anthropic==0.25.0
google-generativeai==0.5.4
openai==1.30.1
httpx==0.27.0
aiohttp==3.9.5

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
tenacity==8.3.0
structlog==24.1.0
rich==13.7.1

# Observability
opentelemetry-sdk==1.24.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
